Specify the parameters of the file metadata and return a list, JSON string, or write to a file
Source: R/07-misc.R

Specify the parameters of the file metadata and return a list, JSON string, or write to a file.

misc_make_metadata(output = c("list", "json", "metafile"), destfile = NULL,
  name = NULL, file_type = c("text", "binary", "fasta", "csfasta", "fastq",
  "qual", "xsq", "sff", "bam", "bam_index", "illumina_export", "vcf", "sam",
  "bed", "archive", "juncs", "gtf", "gff", "enlis_genome"),
  qual_scale = c("sanger", "illumina13", "illumina15", "illumina18",
  "solexa"), seq_tech = c("454", "Helicos", "Illumina", "Solid",
  "IonTorrent"), sample = NULL, library = NULL, platform_unit = NULL,
  paired_end = NULL)

Arguments

output
Output format, could be 'list', 'json', or 'metafile'.

destfile
Filename to write to. Must be specified when output = 'metafile'.

name
File name.

file_type
File type. This metadata parameter is mandatory for each file.

qual_scale
Quality scale encoding. For FASTQ files, you must either specify the quality score encoding sch which contains the FASTQ quality scale detector wrapper. In that case, you can specify the quality score encoding scheme by setting qual_scale inside the pipeline. For BAM files, this value should always be 'sanger'.

seq_tech
Sequencing technology. The seq_tech parameter allows you to specify the sequencing technology used. This metadata parameter is only required by some the tools and pipelines; however, it is strongly recommended that you set it whenever possible, unless you are certain that your pipeline will work without it.

sample
Sample ID. You can use the sample parameter to specify the sample identifier. The value supplied in this field will be written to the read group tag (@RG:SM) in SAM/BAM files generated from reads with the specified Sample ID. AddOrReplaceReadGroups will use this parameter as the value for the read group tag in a SAM/BAM file.

library
Library. You can set the library for the read using the library parameter. The value supplied in this field will be written to the read group tag (@RG:LB) in SAM/BAM files generated from reads with the specified Library ID. AddOrReplaceReadGroups will use this parameter as the value for the read group tag in a SAM/BAM file.

platform_unit
Platform unit. You can set the platform unit (e.g. lane for Illumina, or slide for SOLiD) using the platform_unit parameter. The value supplied in this field will be written to the read group tag (@RG:PU) in SAM/BAM files generated from the reads with the specified Platform Unit. AddOrReplaceReadGroups will use this parameter as the value for the read group tag of a SAM/BAM file.

paired_end
Paired end. With paired-end reads, this parameter indicates if the read file is left end (1) or right end (2). For SOLiD CSFASTA files, paired end files 1 and 2 correspond to R3 and F3 files, respectively.

Value

list, JSON string, or a file.

Details

For more information about file metadata, please check the File Metadata Documentation: https://developer.sbgenomics.com/platform/metadata.

References

https://developer.sbgenomics.com/platform/metadata

Examples

# NOT RUN {
destfile = '~/c.elegans_chr2_test.fastq.meta'
# }# NOT RUN {
misc_make_metadata(output = 'metafile',
            destfile = destfile,
            name = 'c.elegans_chr2_test.fastq',
            file_type = 'fastq', qual_scale = 'illumina13',
            seq_tech = 'Illumina')
# }