Add a pipeline to a specified project
Source: R/05-pipelines.R

Add a pipeline to a specified project. You can use this function to add a pipeline from your other project or a public pipeline to a project.

pipeline_add(auth_token = NULL, project_id_to = NULL,
  project_id_from = NULL, pipeline_id = NULL, revision = NULL, ...)

Arguments

auth_token
auth token

project_id_to
ID of a project you to copy pipeline into.

project_id_from
ID of the project you wish to add from. Specify values such as "my" to specify a pipeline from "My Pipelines" section or omit for a public pipeline, respectively.

pipeline_id
ID of the pipeline you wish to add to project.

revision
Revision of the pipeline you wish to add to the project.

...
parameters passed to sbgapi function

Value

parsed list of the returned json

Examples

# NOT RUN {
token = '58aeb140-1970-0130-6386-001f5b34aa78'
# }# NOT RUN {
req = pipeline_add(token,
            project_id_to = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
            project_id_from = 'f0eb447f-3511-4b28-9253-eba96191d432',
            pipeline_id = '53452130d79f0049c0c94441')
# }